Fix file provider anchor check
This code was modeled after how GitHub is parsing the Markdown. But this
is broken in the sense that our parser may and probably don't follow
the same rules from GitHub's parser.
The change is to parse the fragment using the same library that is used
to process the markdown into HTML. With this we can hit the links
correctly.